Moved files related to the beams data structure and interpolation into beams folder
