+++ src/restruct_poc/AssembleResidualVector.hpp
@@ -4,8 +4,8 @@
 #include <Kokkos_Profiling_ScopedRegion.hpp>
 
 #include "IntegrateResidualVector.hpp"
-#include "beams.hpp"
-#include "types.hpp"
+#include "src/restruct_poc/beams/Beams.hpp"
+#include "src/restruct_poc/types.hpp"
 
 namespace openturbine {
