+++ src/restruct_poc/IntegrateElasticStiffnessMatrix.hpp
@@ -2,7 +2,7 @@
 
 #include <Kokkos_Core.hpp>
 
-#include "beams.hpp"
+#include "beams/Beams.hpp"
 #include "types.hpp"
 
 namespace openturbine {
