+++ src/restruct_poc/beams/BeamNode.hpp
