+++ src/restruct_poc/beams/Beams.hpp
@@ -4,7 +4,7 @@
 
 #include <KokkosBlas.hpp>
 
-#include "types.hpp"
+#include "src/restruct_poc/types.hpp"
 
 namespace openturbine {
