+++ src/restruct_poc/beams/InterpolateQPAcceleration.hpp
@@ -2,7 +2,8 @@
 
 #include <Kokkos_Core.hpp>
 
-#include "types.hpp"
+#include "Beams.hpp"
+#include "src/restruct_poc/types.hpp"
 
 namespace openturbine {
