+++ src/restruct_poc/beams/SetNodeStateIndices.hpp
