+++ src/restruct_poc/solver/Step.hpp
@@ -13,7 +13,7 @@
 #include "UpdateStatePrediction.hpp"
 
 #include "src/restruct_poc/UpdateState.hpp"
-#include "src/restruct_poc/beams.hpp"
+#include "src/restruct_poc/beams/Beams.hpp"
 
 namespace openturbine {
