+++ tests/unit_tests/restruct_poc/test_beams.cpp
@@ -7,13 +7,13 @@
 #include "src/restruct_poc/AssembleInertialStiffnessMatrix.hpp"
 #include "src/restruct_poc/AssembleMassMatrix.hpp"
 #include "src/restruct_poc/AssembleResidualVector.hpp"
-#include "src/restruct_poc/BeamElement.hpp"
-#include "src/restruct_poc/BeamNode.hpp"
-#include "src/restruct_poc/BeamSection.hpp"
-#include "src/restruct_poc/BeamsInput.hpp"
-#include "src/restruct_poc/CreateBeams.hpp"
+#include "src/restruct_poc/beams/BeamElement.hpp"
+#include "src/restruct_poc/beams/BeamNode.hpp"
+#include "src/restruct_poc/beams/BeamSection.hpp"
+#include "src/restruct_poc/beams/BeamsInput.hpp"
+#include "src/restruct_poc/beams/CreateBeams.hpp"
 #include "src/restruct_poc/UpdateState.hpp"
-#include "src/restruct_poc/beams.hpp"
+#include "src/restruct_poc/beams/Beams.hpp"
 #include "src/restruct_poc/solver/State.hpp"
 #include "tests/unit_tests/gen_alpha_poc/test_utilities.h"
