+++ tests/unit_tests/restruct_poc/test_cantilever_beam.cpp
@@ -4,12 +4,12 @@
 
 #include <gtest/gtest.h>
 
-#include "src/restruct_poc/BeamElement.hpp"
-#include "src/restruct_poc/BeamNode.hpp"
-#include "src/restruct_poc/BeamSection.hpp"
-#include "src/restruct_poc/BeamsInput.hpp"
-#include "src/restruct_poc/CreateBeams.hpp"
-#include "src/restruct_poc/beams.hpp"
+#include "src/restruct_poc/beams/BeamElement.hpp"
+#include "src/restruct_poc/beams/BeamNode.hpp"
+#include "src/restruct_poc/beams/BeamSection.hpp"
+#include "src/restruct_poc/beams/BeamsInput.hpp"
+#include "src/restruct_poc/beams/CreateBeams.hpp"
+#include "src/restruct_poc/beams/Beams.hpp"
 #include "src/restruct_poc/solver/InitializeConstraints.hpp"
 #include "src/restruct_poc/solver/Solver.hpp"
 #include "src/restruct_poc/solver/Step.hpp"
